requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.9.3
pyTelegramBotAPI==4.14.0
python-decouple==3.8
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
translate==3.6.1
urllib3==2.0.7
wikipedia==1.4.0
